third_party/blink/web_tests/platform/win7/external/wpt/html/dom/documents/resource-metadata-management/document-lastModified-01-expected.txt

This is a testharness.js-based test.
FAIL Date returned by lastModified is current at page load assert_approx_equals: expected 1566168278 +/- 2.5 but got 1566171877
PASS Date returned by lastModified is in the form "MM/DD/YYYY hh:mm:ss".
FAIL Date returned by lastModified is in the user's local time zone. assert_approx_equals: Hours and minutes should match local time. expected 53077 +/- 2 but got 56677
FAIL Date returned by lastModified is current after timeout. assert_approx_equals: (initial value was 1566171877) expected 1566168282 +/- 2.5 but got 1566171881
Harness: the test ran to completion.
